+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### 2.1.0 (Roadmap)
 - Minimum supported version of Puppet bumped to 5.0.
 - Minimum Ruby version bumped to 2.4.
-- Official support for Puppet version 7.
+- Official support for Puppet 7, Rubocop 1, and Reek 6.
 - Fix Puppet message string transform conditionals.
 
 ### 2.0.1